django-vendas/requirements.txt
Django==4.2.7
mysqlclient==2.2.0
python-decouple==3.8
requests==2.31.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
